server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
transformers==4.35.2
sentence-transformers==2.2.2
pymilvus==2.3.3
python-multipart==0.0.6
aiohttp==3.9.1
nltk==3.8.1
python-dotenv==1.0.0
numpy==1.26.4  # Updated to a version compatible with Python 3.12
torch==2.2.1  # Updated to a compatible version